modularize and some modification
