+++ planning/planning_tuning_tools/scripts/astar_tuning_tools/.gitignore
@@ -1,2 +1,3 @@
 result/
 costmap/
+*.egg-info/
