+++ planning/planning_tuning_tools/scripts/astar_tuning_tools/setup.py
@@ -1,0 +1,4 @@
+from setuptools import find_packages
+from setuptools import setup
+
+setup(name="astar_tuning_tools", version="1.0.0", packages=find_packages())
